ioteca_api/requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
djangorestframework==3.3.2
docutils==0.12
wheel==0.24.0
